fix(console): refactor back to hook but with params
